Remove ' printf flag and change Slayers
Add longfmt and uintfmt to format numbers instead of ' printf flag.
Change Slayer.damage from unsigned long to unsigned int.

+++ src/nolan.h
@@ -48,7 +48,6 @@ enum {
 	UPDATE,
 	USERID,
 };
-/* ALL FIELDS MUST HAVE THE SAME SIZE */
 typedef struct {
 	char *name;
 	char *kingdom;
@@ -79,7 +78,7 @@ typedef struct {
 /* this is for raids */
 typedef struct {
 	char *name;
-	unsigned long damage;
+	unsigned int damage;
 	int found_in_file;
 } Slayer;
 
@@ -118,7 +117,7 @@ void on_stats_admin_interaction(struct discord *client,
 
 /* raids.c */
 void on_raids(struct discord *client, const struct discord_message *event);
-unsigned long parse_file(char *fname, char *username, size_t namelen);
+unsigned int parse_file(char *fname, char *username, size_t namelen);
 
 /* roles.c */
 void update_roles(struct discord *client, Player *player);
